Move some environments code to its own file
Move out the functions for setting up environments and for calculating
environment hashes.
